+++ spec/ddtrace/context/flush/partial_spec.rb
@@ -36,7 +36,7 @@ RSpec.describe Datadog::Context::Flush::Partial do
               finished_spans
             end
 
-            allow(context).to receive(:configure_root_span).with(finished_spans[0])
+            allow(context).to receive(:annotate_for_flush).with(finished_spans[0])
           end
 
           it 'returns finished spans' do
@@ -45,7 +45,7 @@ RSpec.describe Datadog::Context::Flush::Partial do
 
           it 'apply root span settings to first span' do
             subject
-            expect(context).to have_received(:configure_root_span).with(finished_spans[0])
+            expect(context).to have_received(:annotate_for_flush).with(finished_spans[0])
           end
         end
