+++ spec/ddtrace/context_spec.rb
@@ -23,7 +23,7 @@ RSpec.describe Datadog::Context do
         span.sampled = sampled
         context.add_span(span)
 
-        allow(context).to receive(:configure_root_span)
+        allow(context).to receive(:annotate_for_flush)
       end
 
       context 'unfinished' do
@@ -31,7 +31,7 @@ RSpec.describe Datadog::Context do
 
         it 'does not configure unfinished root span' do
           subject
-          expect(context).to_not have_received(:configure_root_span)
+          expect(context).to_not have_received(:annotate_for_flush)
         end
       end
 
@@ -44,7 +44,7 @@ RSpec.describe Datadog::Context do
 
         it 'configures root span' do
           subject
-          expect(context).to have_received(:configure_root_span)
+          expect(context).to have_received(:annotate_for_flush)
         end
       end
     end
@@ -106,7 +106,7 @@ RSpec.describe Datadog::Context do
   end
 
   describe '#delete_span_if' do
-    subject(:configure_root_span) { context.delete_span_if(&block) }
+    subject(:annotate_for_flush) { context.delete_span_if(&block) }
 
     let(:remaining_span) { Datadog::Span.new(tracer, 'remaining', context: context).tap(&:finish) }
     let(:deleted_span) { Datadog::Span.new(tracer, 'deleted', context: context).tap(&:finish) }
@@ -132,8 +132,8 @@ RSpec.describe Datadog::Context do
     end
   end
 
-  describe '#configure_root_span' do
-    subject(:configure_root_span) { context.configure_root_span }
+  describe '#annotate_for_flush' do
+    subject(:annotate_for_flush) { context.annotate_for_flush }
     let(:root_span) { Datadog::Span.new(nil, 'dummy') }
 
     let(:options) { { origin: origin, sampled: sampled, sampling_priority: sampling_priority } }
